Add optional `condition` to `not_null_proportion` test
The dbt core `not_null` test allows a `where` parameter and the dbt
utils `expression_is_true` allows a `condition` parameter, both of which
allow the test to operate on a subset of rows.

This adds a similar functoinality to the `not_null_proportion` test,
sticking with the `condition` convention from `expression_is_true.