Add model conditions to equality test
This adds the following parameters to generic_tests/equality.sql:

- model_condition
- compare_model_condition

It allows one to filter out rows from either model before comparing the
remaining rows in the models as usual.

The rationale and usage is very similar to the relationships_where test.

Note that model_condition is introduced, because the normal 'where'
clause can't be used because it means the given "model" is no longer a
relation, and so the dbt_utils._is_relation check will fail.