equality test: don't always require a Relation
The equality test only needs a Relation when it needs to examine the
model schema using the adapter.  If the user provides all column names
to compare, and does not use the 'precision' option, then no Relation
is actually needed.  In this case, an arbitrary subquery can be passed
instead.  For example, the get_where_subquery can be used.